CosmosDB emulator tests
