+++ workers/tests/emulator_tests/cosmosdb_functions/cosmosdb_functions_stein/function_app.py
