+++ workers/tests/emulator_tests/cosmosdb_functions/cosmosdb_functions_stein/generic/function_app.py
