+++ workers/tests/emulator_tests/cosmosdb_functions/cosmosdb_trigger/__init__.py
