+++ workers/tests/emulator_tests/cosmosdb_functions/cosmosdb_v3_functions_stein/function_app.py
