+++ workers/tests/emulator_tests/cosmosdb_functions/get_cosmosdb_triggered/function.json
